sysclk-lwla: Cut down on size_t overuse
Do not use size_t for values whose width is defined by the device,
not the host. Also don't use size_t for simple indices with known
small range, unless type compatibility considerations apply.

+++ src/hardware/sysclk-lwla/api.c
@@ -372,7 +372,7 @@ static int config_get(uint32_t key, GVariant **data, const struct sr_dev_inst *s
 		      const struct sr_channel_group *cg)
 {
 	struct dev_context *devc;
-	size_t idx;
+	unsigned int idx;
 
 	(void)cg;
